Added namespace declaration for branding namespace
I added some code to add the namespace declaration for the $branding
namespace prefix that is used for IDs. This will clear up validation
errors related to that namespace being undefined. I create a new variable
called "brandingNamespace" to hold the namespace itself but for now
just set it to "ttDemo", same as branding.