Fix validation issue in STIX_Header
In the old code, the stixCommon:Time element was directly under
STIX_Header. Instead, it needs to be nested under stix:Information_Source.

There's unfortunately still a validation issue in the time code, however.
The timezone offset is being formatted incorrectly per the XML spec and so
even with this fix unless you happen to be UTC the document will still not
be valid STIX. I'll submit an issue for this, it's beyond my javascript
abilities.